// models/booktag.ts
export enum BookTag {
  Drama = "Drama",
  Comedy = "Comedy",
  Horror = "Horror",
  Thriller = "Thriller",
  Romance = "Romance",
  Mystery = "Mystery",
  Fantasy = "Fantasy",
  ScienceFiction = "Science Fiction",
  Biography = "Biography",
  SelfHelp = "Self Help",
  Health = "Health",
  Humor = "Humor",
  History = "History",
  Poetry = "Poetry",
  Essay = "Essay",
  Memoir = "Memoir",
  Autobiography = "Autobiography",
  Business = "Business",
  Travel = "Travel",
  Journalism = "Journalism",
  Art = "Art",
  Literature = "Literature"
}